Handle Transform objects by looking at their left hand side column
If the Transform has a 'key_name' attribute, it's probably a KeyTransform
object; in that case the target field is the column name and the JSON
key that we're searching for, concatenated together with a '.'.

If not, we just use the Transform's left hand side for the column and
field values.